styling and clean up components of unnecessary stylings
