+++ client/src/components/ChangePassword.js
@@ -1,17 +1,14 @@
 import React from 'react'
 import Button from '@mui/material/Button';
-import CssBaseline from '@mui/material/CssBaseline';
 import TextField from '@mui/material/TextField';
 import Typography from '@mui/material/Typography';
 import Container from '@mui/material/Container';
 import Box from '@mui/material/Box';
-import { createTheme, ThemeProvider } from '@mui/material/styles';
 import { useState } from 'react'
 import axios from 'axios'
 import {useParams, useNavigate} from 'react-router-dom'
 
 export default function ChangePassword() {
-    const theme = createTheme();
     const navigate = useNavigate()
     const {token} = useParams()
     const [password, setPassword] = useState('')
@@ -32,10 +29,7 @@ export default function ChangePassword() {
     }
 
   return (
-    <div>
-    <ThemeProvider theme={theme}>
     <Container component="main" maxWidth="xs">
-    <CssBaseline />
     <Box
         sx={{
         marginTop: 8,
@@ -89,7 +83,5 @@ export default function ChangePassword() {
         </Box>
     </Box>
     </Container>
-    </ThemeProvider>
-    </div>
   )
 }
