+++ client/src/components/ForgotPassword.js
@@ -1,16 +1,14 @@
 import React from 'react'
 import Button from '@mui/material/Button';
-import CssBaseline from '@mui/material/CssBaseline';
 import TextField from '@mui/material/TextField';
 import Typography from '@mui/material/Typography';
 import Container from '@mui/material/Container';
 import Box from '@mui/material/Box';
-import { createTheme, ThemeProvider } from '@mui/material/styles';
 import { useState } from 'react'
 import axios from 'axios'
 
 export default function ForgotPassword() {
-    const theme = createTheme();
+
     const [data, setData] = useState('')
 
     const handleSubmit = async (event) => {
@@ -22,10 +20,7 @@ export default function ForgotPassword() {
     }
 
   return (
-    <div>
-<ThemeProvider theme={theme}>
 <Container component="main" maxWidth="xs">
-  <CssBaseline />
   <Box
     sx={{
       marginTop: 8,
@@ -63,7 +58,5 @@ export default function ForgotPassword() {
     </Box>
     </Box>
 </Container>
-</ThemeProvider>
-    </div>
   )
 }
